// projects/ccf-database/src/lib/interfaces.d.ts
export interface OntologyTreeNode {
    /** Identifier / IRI */
    '@id': string;
    /** JSON-LD Type */
    '@type': 'OntologyTreeNode';
    /** Identifier / IRI */
    id: string;
    /** RDFS label */
    label: string;
    /** Parent ontology node */
    parent: string;
    /** Child ontology nodes */
    children: string[];
    /** Synonym labels */
    synonymLabels: string[];
}
/** Ontology tree model */
export interface OntologyTreeModel {
    /** id of the root node of the ontology */
    root: string;
    /** Mapping from id/IRI to ontology node instance */
    nodes: {
        [id: string]: OntologyTreeNode;
    };
}